requirements.txt
dotenv
os
pip-system-certs
psycopg2-binary
transformers 
torch 